backend/requirements.txt
#
# StarCourier Web - Backend Requirements (Python 3.13 compatible)
#
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
anyio==4.6.2.post1
sniffio==1.3.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
click==8.1.7
rich==13.9.4

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0

# Authentication
bcrypt==4.2.1
PyJWT==2.10.1
email-validator==2.2.0

# Cache (optional)
# redis==5.2.0
